chore: added JSDocs to hooks, improved context check
